Move the api methods from changeset_comments_controller into the api namespaced controller
